requirements.txt
sqlalchemy~=2.0.41
psycopg2-binary
python-dotenv~=1.1.0

ttkbootstrap~=1.13.11
pyodbc~=5.1.0